update: refactor to a view file
